requirements.txt
loadit
torch
torchvision
wandb
tqdm
datasets
transformers
wandb
hydra-core
einops
tqdm
# we install jax and related packages last because otherwise pytorch will overwrite the CUDA installation with
# an older one and jax will complain that it cannot use the GPU
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html "jax[cuda12_pip]" 
optax
equinox
jaxamp